feat: git credential details modal seperated
